[sailfish-browser] Fix error reading tab thumbnails from cache. Fixes JB#56298 OMP#JOLLA-533
QImage (Qt 5.6) writes directly to the target file when saving, so if an
Image item happens to try and read from it at the same file then that
may fail because the file is incomplete.

The concurrent read/write may happen because an existing tab is being
updated, or because saveToFile() was dispatched to a thread and the user
could have switched tabs in the interim leading to the data being
written to the thumbnail of a different tab.